rewrote the css with felxbox
Reworte the entire css with felxbox alignment removing the hard to use float positioning. Thus removed unwanted classes from the html page.

Seperated styles for mobile and desktop with seperate stylesheets.

Hid the support menu system because it is not nessary for this design and will be removed in subsequent iterations.